pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gex-monitor"
version = "2.0.0"
description = "0DTE Gamma Exposure Monitor"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "ib_insync>=0.9.86",
    "pandas>=2.0",
    "numpy>=1.24",
    "dash>=2.14",
    "plotly>=5.18",
    "pyarrow>=14.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "exchange-calendars>=4.2",
    "flask>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
]

[project.scripts]
gex-monitor = "gex_monitor.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "W"]
